nixos-rebuild-ng: move models.Ssh to process.Remote
